services: Use launcchtl enable+bootstrap, not deprecated load -w
"launchctl load" has been deprecated since 10.10/Yosemite, if not
before.  There are at least two good reasons not to use it beyond
that:
  1. It has bad error reporting
  2. It behaves differently as root vs. user

Instead, we move to "launchctl enable" and "launchctl bootstrap", both
of which are non-deprecated. We mimic the "load -w" behavior of
determining the domain target based on whether we are root.  Perhaps
in a future restructuring that should change.

NB: Under 10.10, the mirror image of "launchctl boostrap" is
unimplemented, namely "launchctl unbootstrap" (or, as it's called in
10.11, "launchctl bootout"). So we continue to use "launchctl unload
-w" to remove services. This is fine and does not cause any problems.