Added formula_wrapper#load_service method
This helps load Homebrew::Service class and returns the @formula.service object.

The main reason for having a separate method here is to make obvious that the
class needs to be loaded using the require_relative command and that the @formula.service
class should not be accessed directly. It also helps avoid nameclashes between the
Homebrew::Service class and Service module used by this subcommand.

All calls to formula.service have been replaced with load_service.

I also updated the tests to reflect the new calling patterns.
1. #service? is used instead of formula#service?
2. #service? gets called twice in #to_hash because of the call to #loaded?
3. #load_service gets called twice in #to_hash because of the call to #loaded?