Add `--sudo-service-user` flag
This flag allows setting the `UserName` for `launchd` plists when
running as `root`. This means that `launchd` will drop privileges to the
relevant user when starting the service.

This is significantly more secure than running services as `root` and
enables a reasonable way to run systemwide services as unprivileged
users.