Fix style issues
`shellcheck` is reporting SC2046 (Quote this to prevent word
splitting) and SC2086 (Double quote to prevent globbing and word
splitting) issues in the `tests.yml` workflow. Adding quotes in
affected areas resolves these warning/info messages but I haven't
tested these changes, so take this with a grain of salt.