Add `--max-wait` flag for `stop` subcommand
We currently only have a choice between passing `--no-wait` or omitting
it. In the former, we don't wait at all for services to stop. In the
latter, we wait as long as it takes (which could be forever, if some
services have frozen for some reason).

This enables a middle ground where we wait only up to a given amount of
time until we're no longer willing to wait for a service to stop.