Fix missing `max_wait` argument in `::Service::Commands::Stop.run` call
### Summary
This PR fixes an issue in the `::Service::Commands::Stop.run` method call, where `max_wait` was referenced without being assigned a value.

### Changes
- Updated the `max_wait:` argument to `max_wait: max_wait` to correctly pass the intended `max_wait` variable.

### Context
The lack of a value assignment to `max_wait` resulted in an error when invoking the stop command. This change ensures that `max_wait` is now passed correctly as an argument.